requirements.txt
certifi==2021.5.30
cycler==0.10.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
numpy==1.19.5
olefile==0.46
opencv-python==4.5.4.60
Pillow @ file:///C:/ci/pillow_1617386325740/work
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
wincertstore==0.2
